CHORE: mypage navbar css
