+++ src/components/mypage/MyNavbar.js
@@ -1,0 +1,29 @@
+import styles from "./MyNavbar.module.css";
+import {NavLink} from "react-router-dom";
+
+function MyNavbar() {
+    const activeStyle={
+        color: 'blue',
+        textDecoration: "none"
+    };
+    const unactiveStyle={
+        color: 'black',
+        textDecoration: "none"
+    };
+
+    return(
+        <div>
+            <div className={styles.head}>MY PAGE</div>
+            <ul className={styles.nav}>
+                <li><NavLink to="/MyPage" style={({ isActive }) => (isActive ? activeStyle : unactiveStyle)}>프로필 수정</NavLink></li>
+                <li><NavLink to="/MyMentoApply" style={({ isActive }) => (isActive ? activeStyle : unactiveStyle)}>멘토링 신청</NavLink></li>
+                <li><NavLink to="/MyMento" style={({ isActive }) => (isActive ? activeStyle : unactiveStyle)}>멘토링 내역</NavLink></li>
+                <li><NavLink to="/MyStudy" style={({ isActive }) => (isActive ? activeStyle : unactiveStyle)}>스터디 관리</NavLink></li>
+                <li><NavLink to="/MyScrap" style={({ isActive }) => (isActive ? activeStyle : unactiveStyle)}>스크랩 내역</NavLink></li>
+                <li><NavLink to="/MyPoint" style={({ isActive }) => (isActive ? activeStyle : unactiveStyle)}>포인트 관리</NavLink></li>
+            </ul>
+        </div>
+    );
+}
+
+export default MyNavbar;
