+++ src/routes/Mypage.module.css
@@ -1,45 +1,17 @@
 .wrap {
-  width: 1300px;
-  height: 1000px;
-  margin: 0 auto;
+  padding: 0 16%;
 }
 
-.head {
-  height: 90px;
-  padding-top: 20px;
-  font-weight: 600;
-  font-size: 36px;
-}
-
-.nav {
-
-  text-align: center;
-  width: 1300px;
-  height: 50px;
-  line-height: 44px;
-  list-style: none;
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  border-top: 2px solid #eee;
-  border-bottom: 2px solid #eee;
-}
-
-.nav li>* {
-  width: 180px;
-  font-size: 24px;
-
-}
-
-.menu {
+/* .menu {
   text-align: right;
   margin-right: 50px;
-}
+} */
 
 .body{
     display: flex;
-    width: 800px;
-    margin: 0 auto;
+    padding: 0 2%;
+    /* width: 800px;
+    margin: 0 auto; */
 }
 .left{
     margin-right: 200px;
